+++ lib/auth/recordingencryption/recordingencryptionv1/service.go
@@ -31,7 +31,7 @@ import (
 	"github.com/gravitational/teleport/lib/auth/summarizer"
 	"github.com/gravitational/teleport/lib/authz"
 	"github.com/gravitational/teleport/lib/events"
-	sessionpostprocessing "github.com/gravitational/teleport/lib/events/postprocessing"
+	sessionpostprocessing "github.com/gravitational/teleport/lib/events/sessionpostprocessing"
 	"github.com/gravitational/teleport/lib/session"
 )
 
@@ -227,9 +227,9 @@ func (s *Service) CompleteUpload(ctx context.Context, req *recordingencryptionv1
 		return &recordingencryptionv1.CompleteUploadResponse{}, nil
 	}
 
-	if err := sessionpostprocessing.SessionPostProcessor(
+	if err := sessionpostprocessing.Process(
 		ctx,
-		sessionpostprocessing.SessionPostProcessorConfig{
+		sessionpostprocessing.Config{
 			SessionEnd:                sessionEnd,
 			SessionID:                 upload.SessionID,
 			SessionSummarizerProvider: s.sessionSummarizerProvider,
